Moved authentication utility functions to /utils/
- Created `authentication.js` in `/utils/`.
- Moved `isPasswordValid` and `generateAuthToken`.
- Wrote documentation for both functions.
